requirements.txt
django>=5.0,<5.1
django-ninja>=1.4.0,<2.0.0
celery>=5.5.0,<6.0.0
django-celery-beat>=2.5.0,<3.0.0
django-elasticsearch-dsl>=8.0,<9.0
redis>=5.2.1,<6.0.0
django-redis>=5.4.0,<6.0.0
uvicorn>=0.34.0,<0.35.0
pyjwt>=2.10.1,<3.0.0
requests>=2.32.3,<3.0.0
dj-database-url>=2.1.0,<3.0.0 